+++ source/extensions/isaacsim.ros2.core/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.2.2] - 2025-11-26
+### Changed
+- PointCloud2 message backend now optionally supports host-pinned buffer
 
 ## [1.2.1] - 2025-11-26
 ### Added